pyproject.toml
[tool.poetry]
name = "sssom"
version = "0.0.0"
description = "Operations on SSSOM mapping tables"
authors = [
    "Chris Mungall <cjmungall@lbl.gov>",
    "Nicolas Matentzoglu <nicolas.matentzoglu@gmail.com>",
    "Harshad Hegde <hhegde@lbl.gov>"
]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
click = ">=8.1.6"
curies = ">=0.7.3"
linkml-runtime = ">=1.5.5"
importlib-metadata = ">=6.8.0"
pandas = ">1.0.3"
pansql = ">=0.0.1"
sssom-schema = ">=0.14.0"
networkx = ">=3.1"
sparqlwrapper = ">=2.0.0"
validators = ">=0.20.0"
deprecation = "^2.1.0"
pyyaml = "^6.0.1"
rdflib = ">=6.0.0"
scipy = {version = "*", extras = ["scipy"]}

[tool.poetry.group.dev.dependencies]
pytest = {version = ">=7.1.2"}
tox = {version = ">=3.25.1"}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = {version = ">=6.1.3"}
sphinx-rtd-theme = {version = ">=1.0.0"}
sphinx-autodoc-typehints = {version = ">=1.2.0"}
sphinx-click = {version = ">=4.3.0"}
myst-parser = {version = ">=0.18.1"}

[tool.poetry.scripts]
sssom = "sssom.cli:main"

[tool.poetry-dynamic-versioning]
enable = false
vcs = "git"
style = "pep440"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310"]

[[tool.mypy.overrides]]
 module = [
     'sssom.sssom_datamodel',
     'sssom.cliquesummary'
 ]
 ignore_errors = true
    

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
include_trailing_comma = true
reverse_relative = true

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
